+++ app/src/main/java/com/yoga/mborasystem/viewmodel/ProdutoViewModel.java
@@ -505,7 +505,7 @@ public class ProdutoViewModel extends AndroidViewModel {
                 }));
     }
 
-    public LiveData<Integer> consultarQuantidadeProduto(long idcategoria) {
+    public LiveData<Long> consultarQuantidadeProduto(long idcategoria) {
         return produtoRepository.getQuantidadeProduto(idcategoria);
     }
